+++ src/card/cards/counted_bodies.rs
@@ -7,3 +7,5 @@ pub const PLAGUE_RATS: CardDefinitionId = CardDefinitionId(1466);
 pub const KELDON_WARLORD: CardDefinitionId = CardDefinitionId(1467);
 pub const GAEAS_AVENGER: CardDefinitionId = CardDefinitionId(1468);
 pub const DAKKON_BLACKBLADE: CardDefinitionId = CardDefinitionId(1469);
+pub const PEOPLE_OF_THE_WOODS: CardDefinitionId = CardDefinitionId(1470);
+pub const TETSUO_UMEZAWA: CardDefinitionId = CardDefinitionId(1471);