+++ src/card/model/presentation.rs
@@ -78,6 +78,7 @@ fn object_predicate_implies(predicate: ObjectPredicateDef, expected: ObjectPredi
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -114,6 +115,7 @@ fn predicate_color(predicate: ObjectPredicateDef) -> Option<ManaColor> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -152,6 +154,7 @@ fn predicate_color_count(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -190,6 +193,7 @@ fn predicate_subtype(predicate: ObjectPredicateDef) -> Option<&'static str> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -232,6 +236,7 @@ fn predicate_negated_subtype(predicate: ObjectPredicateDef) -> Option<&'static s
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -276,6 +281,7 @@ fn predicate_power_at_least(predicate: ObjectPredicateDef) -> Option<i16> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -312,6 +318,7 @@ fn predicate_mana_value_at_most(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -350,6 +357,7 @@ fn predicate_controller(predicate: ObjectPredicateDef) -> Option<PlayerRelation>
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell
@@ -391,6 +399,7 @@ fn predicate_negates(predicate: ObjectPredicateDef, expected: ObjectPredicateDef
         | ObjectPredicateDef::Token
         | ObjectPredicateDef::Tapped
         | ObjectPredicateDef::Attacking
+        | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::AttackedThisTurn
         | ObjectPredicateDef::HasType(_)
         | ObjectPredicateDef::Spell